+++ conjurapi/config.go
@@ -4,6 +4,9 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
+	"os/user"
+	"path"
+	"runtime"
 	"strings"
 
 	"github.com/cyberark/conjur-api-go/conjurapi/logging"
@@ -115,15 +118,20 @@ func (c *Config) mergeEnv() {
 }
 
 func LoadConfig() Config {
+    usr, err := user.Current()
+	if err != nil {
+		logging.ApiLog.Debugf("Failed to get os.user object")
+		return Config{}
+	}
 	// Default to using ~/.netrc, subsequent configuration can
 	// override it.
-	config := Config{NetRCPath: os.ExpandEnv("$HOME/.netrc")}
+	config := Config{NetRCPath: path.Join(usr.HomeDir, ".netrc")}
 
-	config.mergeYAML("/etc/conjur.conf")
+	config.mergeYAML(path.Join(getSystemPath(), "conjur.conf"))
 
 	conjurrc := os.Getenv("CONJURRC")
 	if conjurrc == "" {
-		conjurrc = os.ExpandEnv("$HOME/.conjurrc")
+		conjurrc = path.Join(usr.HomeDir, ".conjurrc")
 	}
 	config.mergeYAML(conjurrc)
 
@@ -131,4 +139,14 @@ func LoadConfig() Config {
 
 	logging.ApiLog.Debugf("Final config: %+v\n", config)
 	return config
+}
+
+func getSystemPath() string {
+	if runtime.GOOS == "windows" {
+		//No way to use SHGetKnownFolderPath()
+		//Hardcoding should be fine for now since CONJURRC is available
+		return "C:\\windows"
+	} else {
+		return "/etc"
+	}
 }