+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Use a separate logrus logger for the API. Control the destination for messages with the
   environment variable `CONJURAPI_LOG`.
 
+# [0.4.0]
+
+* Add support for `.conjurrc` and `.netrc` in Windows user directories.
+* Add support for `conjur.conf` in Windows system directory.
+
 # [0.3.3]
 
 * Update the tags on `PolicyResponse` so they match the JSON returned by the server.
@@ -49,6 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Initial version
 
 [Unreleased]: https://github.com/cyberark/conjur-api-go/compare/v0.3.3...HEAD
+[0.4.0]: https://github.com/cyberark/conjur-api-go/compare/v0.3.3...v0.4.0
 [0.3.3]: https://github.com/cyberark/conjur-api-go/compare/v0.3.0...v0.3.3
 [0.3.0]: https://github.com/cyberark/conjur-api-go/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/cyberark/conjur-api-go/compare/v0.1.0...v0.2.0